@@ -7,30 +7,34 @@
   "author": "Ivan Trejo <ivanjtrejoc@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "@emailjs/browser": "^3.6.2",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@mui/material": "^5.14.8",
-    "@mui/styled-engine-sc": "^5.14.8",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "classnames": "^2.2.6",
+    "@emailjs/browser": "^4.3.3",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@gsap/react": "^2.1.1",
+    "@material-ui/core": "^4.12.4",
+    "@material-ui/icons": "^4.11.3",
+    "@mui/icons-material": "^5.15.20",
+    "@mui/lab": "^5.0.0-alpha.170",
+    "@mui/material": "^5.15.20",
+    "@mui/styled-engine-sc": "^5.14.12",
+    "@testing-library/jest-dom": "^6.4.6",
+    "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
+    "classnames": "^2.5.1",
+    "gsap": "^3.12.5",
     "ios-inner-height": "^1.1.1",
-    "popmotion": "^8.7.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "popmotion": "^11.0.5",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
-    "react-router-dom": "^5.3.4",
+    "react-router-dom": "^6.24.0",
     "react-scripts": "^5.0.1",
-    "react-snapshot": "^1.0.4",
-    "react-transition-group": "^4.4.1",
+    "react-snapshot": "^1.3.0",
+    "react-transition-group": "^4.4.5",
     "react-typical": "^0.1.3",
-    "styled-components": "^6.0.7",
-    "sweetalert2": "^11.4.10",
-    "three": "^0.125.0",
-    "use-dencrypt-effect": "1.0.1" 
+    "styled-components": "^6.1.11",
+    "sweetalert2": "^11.12.0",
+    "use-dencrypt-effect": "^2.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -38,11 +42,11 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-    "eslintConfig": {
-    "extends":[
+  "eslintConfig": {
+    "extends": [
       "react-app",
-    "react-app/jest"
-    ] 
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -58,7 +62,7 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "emailjs": "^4.0.0"
-    
+    "@types/three": "^0.165.0",
+    "emailjs": "^4.0.3"
   }
 }
