@@ -1,6 +1,6 @@
 import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
-import { SpeedDial, SpeedDialIcon, SpeedDialAction } from "@material-ui/lab";
+import { SpeedDial, SpeedDialIcon, SpeedDialAction } from "@mui/material";
 import Resume from "../../settings/resume.json";
 
 const useStyles = makeStyles((theme) => ({
