@@ -7,21 +7,24 @@ import { FirstName, LastName } from "../../utils/getName";
 
 const useStyles = makeStyles((theme) => ({
   main: {
-    maxWidth: '100vw',
+    maxWidth: "100vw",
     marginTop: "10vw",
-    marginBottom: "auto",
+    marginBottom: "auto"
   },
   heading: {
     marginLeft: theme.spacing(50),
-    "@media (max-width: 768px)": {
-      marginLeft: theme.spacing(10),
+    "@media (max-width: 1440px)": {
+      marginLeft: theme.spacing(40)
     },
+    "@media (max-width: 768px)": {
+      marginLeft: theme.spacing(10)
+    }
   },
   jobs: {
     "@media (max-width: 768px)": {
-      fontSize: '3rem',
-    },
-  },
+      fontSize: "3rem"
+    }
+  }
 }));
 
 export const Content = () => {
@@ -31,11 +34,11 @@ export const Content = () => {
     <Container component="main" className={classes.main} maxWidth="md">
       <div className={classes.heading}>
         <Typography variant="h5" component="h2">
-            <TextDecrypt text={`${FirstName} ${LastName}`} />
+          <TextDecrypt text={`${FirstName} ${LastName}`} />
         </Typography>
         <Typography variant="h1" component="h1" className={classes.jobs}>
-            <TextDecrypt text={`${Resume.basics.job1}  `} />
-            <TextDecrypt text={`${Resume.basics.job2}`} />
+          <TextDecrypt text={`${Resume.basics.job1}  `} />
+          <TextDecrypt text={`${Resume.basics.job2}`} />
         </Typography>
       </div>
     </Container>
