@@ -1,6 +1,9 @@
 import React from "react";
-import { render } from "react-snapshot";
+import { createRoot } from "react-dom/client";
 import { App } from "./app/App";
 import "./index.css";
 
-render(<App />, document.getElementById("root"));
+const container = document.getElementById("root");
+const root = createRoot(container);
+
+root.render(<App />);
